feat: insert books endpoint

+++ interface/embed_interface.go
@@ -2,7 +2,7 @@ package _interface
 
 type (
 	EmbeddingModel interface {
-		EmbedSearchQuery(query string) ([]float64, error)
-		EmbedSearchDocument(text string) ([]float64, error)
+		EmbedSearchQuery(query string) ([]float32, error)
+		EmbedSearchDocument(text string) ([]float32, error)
 	}
 )
